[#278] Split `Codec.Combinators` into more modular way
Resolves #278
